Rename tabular-playground-series-may-2022.ipynb to TBS - Submission 1 - Base model
The code is for a classification task as part of the Tabular Playground Series - May 2022 competition. The base models built are XGBoost and LightGBM. However, both of them have not been tuned. Stratified K-Fold technique with 10 splits has been used to cross-validate while training each model.
XGBoost performed better than LightGBM and the metric used was accuracy. Further improvements will be made to enhance the performance of the model.